refactor: 미사용 DTO 삭제 및 정렬 로직 Entity 테스트 추가
