+++ src/main/java/upbrella/be/store/service/StoreDetailService.java
@@ -76,12 +76,15 @@ public class StoreDetailService {
 
     public SingleStoreResponse createSingleStoreResponse(StoreDetail storeDetail) {
 
-        AllImageUrlResponse imageUrls = AllImageUrlResponse.fromImagesWithSorted(storeDetail.getSortedStoreImages());
-        String thumbnail = storeImageService.createThumbnail(imageUrls.getSingleImageUrlResponses());
+        List<SingleImageUrlResponse> sortedImageUrls = storeDetail.getSortedStoreImages().stream()
+                .map(SingleImageUrlResponse::createImageUrlResponse)
+                .collect(Collectors.toList());
+
+        String thumbnail = storeImageService.createThumbnail(sortedImageUrls);
         Set<BusinessHour> businessHourSet = storeDetail.getStoreMeta().getBusinessHours();
         List<BusinessHour> businessHourList = new ArrayList<>(businessHourSet);
         List<SingleBusinessHourResponse> businessHours = businessHourService.createBusinessHourResponse(businessHourList);
-        return SingleStoreResponse.ofCreateSingleStoreResponse(storeDetail, thumbnail, imageUrls.getSingleImageUrlResponses(), businessHours);
+        return SingleStoreResponse.ofCreateSingleStoreResponse(storeDetail, thumbnail, sortedImageUrls, businessHours);
     }
 
     public AllStoreIntroductionResponse findAllStoreIntroductions() {
@@ -102,8 +105,11 @@ public class StoreDetailService {
 
     private SingleStoreIntroductionResponse createSingleIntroduction(StoreDetail storeDetail) {
 
-        AllImageUrlResponse imageUrls = AllImageUrlResponse.fromImagesWithSorted(storeDetail.getSortedStoreImages());
-        String thumbnail = storeImageService.createThumbnail(imageUrls.getSingleImageUrlResponses());
+        List<SingleImageUrlResponse> sortedImageUrls = storeDetail.getSortedStoreImages().stream()
+                .map(SingleImageUrlResponse::createImageUrlResponse)
+                .collect(Collectors.toList());
+
+        String thumbnail = storeImageService.createThumbnail(sortedImageUrls);
         StoreMeta storeMeta = storeDetail.getStoreMeta();
         return SingleStoreIntroductionResponse.of(storeDetail.getId(), thumbnail, storeMeta.getName(), storeMeta.getCategory());
     }
